% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exNWchecks.R
\docType{data}
\name{exNWattempts}
\alias{exNWattempts}
\title{Sample NestWatch Attempts Dataframe for Examples}
\format{
A dataframe with 10 observations of 32 variables. See <https://doi.org/10.1002/ecy.4230> for details of variables.
}
\usage{
exNWattempts
}
\description{
A random set of 10 nesting attempts from the NewstWatch database to use used in running R documentation exmaples.
}
\keyword{datasets}
