% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exNWattempts.R
\docType{data}
\name{exNWchecks}
\alias{exNWchecks}
\title{Sample NestWatch Checks Dataframe for Examples}
\format{
A dataframe with 79 observations of 21 variables. See <https://doi.org/10.1002/ecy.4230> for details of variables.
}
\usage{
exNWchecks
}
\description{
A random set of 10 nesting attempts from the NewstWatch database to use used in running R documentation examples.
}
\keyword{datasets}
