% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.cleantaxa.R
\name{nw.cleantaxa}
\alias{nw.cleantaxa}
\title{Preform Cleaning Actions on Species Names & Codes}
\usage{
nw.cleantaxa(
  data,
  spuh = FALSE,
  slash = FALSE,
  hybrid = FALSE,
  rollsubspecies = FALSE,
  output = NULL
)
}
\arguments{
\item{data}{dataframe; A dataframe containing NestWatch data.}

\item{spuh}{logical; Should attempt with "sp." be removed?
 \itemize{
  \item if \code{NULL}, the default, retains identified rows in the output dataframe.
  \item if \code{TRUE}, removes identified rows from the output dataframe.
  \item if \code{FALSE}, retains identified rows in the output dataframe.
}}

\item{slash}{logical; Should attempt with "/" be removed?
 \itemize{
  \item if \code{NULL}, the default, retains identified rows in the output dataframe.
  \item if \code{TRUE}, removes identified rows from the output dataframe.
  \item if \code{FALSE}, retains identified rows in the output dataframe.
}}

\item{hybrid}{logical; Should attempt with "(hybird)" be removed?
 \itemize{
  \item if \code{NULL}, the default, retains identified rows in the output dataframe.
  \item if \code{TRUE}, removes identified rows from the output dataframe.
  \item if \code{FALSE}, retains identified rows in the output dataframe.
}}

\item{rollsubspecies}{logical; Should attempt with Species.Name of the form "species (subspecies/form) be rolled up to the species level?
Uses the current eBird taxonomy as in \code{\link[auk:ebird_taxonomy]{auk::ebird_taxonomy}}.
 \itemize{
  \item if \code{NULL}, the default, retains original Species.Names and Species.Codes.
  \item if \code{TRUE}, preforms taxanomic roll up for all attempts with subspecies/form designation and updates Species.Code.
  \item if \code{FALSE}, retains original Species.Names and Species.Codes.
}}

\item{output}{character; An optional character vector to custom name the output dataframe}
}
\value{
a dataframe
}
\description{
This function can optionally remove attempts for species identified as hybrids, "spuhs", or "slashes" and  preform
a "taxonomic roll up" of subspecies to the species level.
}
\details{
The NestWatch platform uses eBird taxonomy to help users auto-complete species names. This allows users to select any
available eBird taxonomic unit, including spuhs (ie. "chickadee sp."), slashes (ie. "Tree/Violet-green Swallow"),
and subspecies/forms (ie."House Wren (northern)"). In some cases, the user may decide to omit spuhs and slashes because species
is not known with certainty. Or the user may want to analyze all subspecies/forms of a species together if these differentiations
are not biologically relevant to the analysis. This function will update both the Common Name and Species Code of the attempt
according to the current version of eBird taxonomy in \code{\link[auk:ebird_taxonomy]{auk::ebird_taxonomy}}.
}
\examples{
# data with species name variations
df <- data.frame(Attempt.ID = c(1, 1, 2, 3, 4),
                 Species.Name = c("House Wren (northern)", "House Wren (northern)",
                                 "chickadee sp.", "Tree/Violet-green Swallow",
                                 "Mallard x American Black Duck (hybrid)"),
                Species.Code = c("houwre1", "houwre1", "chicka1", "y00701", "x00004"))

# Remove just ".sp" attempts
nw.cleantaxa(df, spuh = TRUE)

# Remove "/" and hybrid attempts, roll up subspecies to species
nw.cleantaxa(df, slash = TRUE, hybrid = FALSE, rollsubspecies = TRUE)
}
\seealso{
\code{\link[auk:ebird_taxonomy]{auk::ebird_taxonomy}}
}
