% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.estclutchsize.R
\name{nw.estclutchsize}
\alias{nw.estclutchsize}
\title{Estimate Missing Summary Clutch Size Values in NestWatch Data}
\usage{
nw.estclutchsize(data, output = NULL)
}
\arguments{
\item{data}{dataframe; A dataframe containing merged NestWatch attempts and visits data.}

\item{output}{character; An optional character vector to rename the resulting output.}
}
\value{
dataframe; A dataframe with an additional binary column \code{Clutch.Szie.Estimated} where \code{1} denotes the clutch size was
estimated by the function and \code{0} denotes the clutch size was provided by the participant.
}
\description{
Clutch size is an important value for estimating nest dates and determining the successfulness of a nest.
NestWatch participants may forget to enter clutch size into the nest attempt's summary data. But clutch size information may be estimated
from the data of individual nest visits. This function used available data from any entered nest visits to estimate maximum host egg count
(clutch size).
}
\details{
This function looks for data contained in several columns to estimate what the clutch size of the nest was. Because NestWatch
participants do not necessarily visit each nest everyday or record all data fields on every visit the data needed to estimate clutch size
may be spread over many columns. Depending on the data available, the certainty of the clutch size estimate may vary, or in some cases,
may not be estimable. For instance, a successful nest with 4 eggs recorded on several checks and 4 fledged young observed would have high
certainty that the clutch size was 4. However, if a nest was observed with 2 eggs on one check and a week later was found predated, the
certainty of the clutch size is not known. In this case the clutch size would be estimated as 2 but may have been higher depending on the
exact (unobserved) failure date.
}
\examples{
# Example simplified NestWatch dataset with only single nest visits:
data <- data.frame(Attempt.ID = c(1, 2, 3, 4, 5, 6),
                   Visit.ID = c(1, 2, 3, 4, 5, 6),
                   Clutch.Size = c(1, NA, NA, NA, NA, NA),
                   Host.Eggs.Count = c(1, 2,  NA, NA, NA, 3),
                   Young.Total = c(1, 2, 3, 3, 4, NA),
                   Live.Host.Young.Count = c(1, 1, 2, 3, 1, 1),
                   Unhatched.Eggs = c(0, 0, NA, NA, 1, 3),
                   Young.Fledged = c(1, 0, NA, NA, 1, 3),
                   Dead.Host.Young.Count = c(0, NA, NA, 1, 0,1))

# Estimate Clutch Sizes if NA
nw.estclutchsize(data = data, output = "out")

out$Clutch.Size   # estimated values
out$Clutch.Size.Estimated   # estaimte binary indicator
}
