% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.mergedata.R
\name{nw.mergedata}
\alias{nw.mergedata}
\title{Merge NestWatch Datasets}
\usage{
nw.mergedata(attempts, checks, output = NULL)
}
\arguments{
\item{attempts}{dataframe; A dataframe containing NestWatch nest attempts data}

\item{checks}{dataframe; A dataframe containing NestWatch nest check data}

\item{output}{character; An optional character vector to custom name the output dataframe}
}
\value{
dataframe; A dataframe of merged nest attempts and checks unified on "Attemp.ID"
}
\description{
Merge NestWatch Datasets
}
\examples{
 \dontrun{
# Create simplified example data
NW.attempts <- data.frame

# Merge downloaded NestWatch datasets
nw.mergedata(NW.attempts, NW.checks)


# Specify the new dataframe name
nw.mergedata(NW.attempts, NW.checks, output = "merged_data")
}
}
