% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wren-quickstart.R
\docType{data}
\name{wren_quickstart}
\alias{wren_quickstart}
\title{Data for NestWatch Workflow Examples}
\format{
A dataframe with 40290 observations of 54 variables. See <https://doi.org/10.1002/ecy.4230> for details of variables.
}
\usage{
wren_quickstart
}
\description{
Data from the NestWatch database (V2), subset to include all nesting data for Carolina ("carwre") and Bewick's ("bewwre") wrens. These are un-cleaned and un-filtered from the database.
}
\keyword{datasets}
