% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.filterphenology.R
\name{nw.filterphenology}
\alias{nw.filterphenology}
\title{Filter Nesting Attempts Based on Expected Nest Phenological Periods}
\usage{
nw.filterphenology(data, sp, mode, phenology, output = NULL)
}
\arguments{
\item{data}{dataframe; A dataframe of merged NestWatch attempt and visits data. Date columns must have class \code{Date}
and Datetime columns must have class \code{POSIXct}.}

\item{sp}{character vector; A character vector containing the species code for one or more species for which to filter the data.}

\item{mode}{\code{"flag"} or \code{"remove"}; A character string defining if the user wants the identified nesting attempts to be
flagged with "FLAGGED" in a new column. Or removed from the dataset.}

\item{phenology}{dataframe; A simple dataframe with one row for each species of interest and the following column structure:
\itemize{
 \item \strong{Species}: Species for which this row of data represents. Must be a 6-letter species code.
 \item \strong{Lay}: The number of days representing the maximum expected laying period for each species.
 \item \strong{Incubation}: The number of days representing the maximum expected incubation period (between clutch complete and hatch) for each species.
 \item \strong{Nestling}: The number of days representing the maximum expected nesting period (hatch and last fledge) for each species.
 \item \strong{Total Nesting Period}: The number of days representing the maximum expected nesting period (spanning first lay to last fledge) for each species.
 }}

\item{output}{character; An optional character string to custom name the output dataframe}
}
\value{
dataframe
}
\description{
To provide additional checks on data quality before analysis, a user may want to filter out nesting attempts which were
recorded to have nesting phenologies of unexpected lengths. This function allows a user to specify specie-specific values for nest phenologies
using expected days in the lay, incubation, nestling, and total nesting periods. By defining these allowable time frames a user can flag or
remove "run-on" nest attempts or nests which do not match the specie's life history.
}
\examples{
\dontrun{
x <- 1+2
}
}
