% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.mergedata.R
\name{nw.mergedata}
\alias{nw.mergedata}
\title{Merge NestWatch Datasets}
\usage{
nw.mergedata(attempts, checks, output = NULL)
}
\arguments{
\item{attempts}{dataframe; A dataframe containing NestWatch nest attempts data}

\item{checks}{dataframe; A dataframe containing NestWatch nest check data}

\item{output}{character; An optional character vector to custom name the output dataframe}
}
\value{
dataframe; A dataframe of merged nest attempts and checks unified on "Attempt.ID"
}
\description{
Merge NestWatch Datasets
}
\examples{

# Load example NestWatch data
ex.attempts <- nestwatchR::exNWattempts
ex.checks <- nestwatchR::exNWchecks

# Merge example NestWatch datasets
nw.mergedata(ex.attempts, ex.checks)

\dontrun{
# Optionally specify the new dataframe name
nw.mergedata(ex.attempts, ex.checks, output = "merged_data")
}
}
